Import all of static/app into "sentry" dll
