--- src/sentry/static/sentry/app/components/index.js
@@ -1,5 +1,0 @@
-require.context(
-  '.', // context folder
-  true, // include subdirectories
-  /.*\.jsx?/ // js or jsx files
-);
